+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Table of Contents]
 - [Unreleased](#unreleased)
+- [0.4.1](#040---2021-06-13)
 - [0.4.0](#040---2021-06-07)
 - [0.3.1](#031---2021-05-28)
 - [0.3.0](#030---2021-05-26)
@@ -12,13 +13,23 @@
 
 ## [Unreleased][]
 ### Added
-- Integration: Listeners to share tracked playlist path with other panels (use 'window.NotifyOthers('Playlist manager: playlistPath', null)'). Used on [Playlist-Tools-SMP](https://github.com/regorxxx/Playlist-Tools-SMP) to use tracked playlist files as source for pools. i.e. Playlist A would match first a playlist within foobar with same name, then a playlist file with matching '#PLAYLIST:Playlist A' tag and finally any playlist file named 'Playlist A.m3u8'. Autoplaylists are excluded (use queries instead) and fpl files too. This feature allows to use virtual playlists as containers, where you can easily collect tracks (since Playlist Manager allows to send tracks directly to a file without loading it) to be used later on pools without polluting the UI with tons of dummy playlists.
 ### Changed
-- Portable: first time properties use now relative paths on profile folder for portable installations (>= 1.6).
-- Integration: Moved some functions to 'playlist_manager_helpers.js' to easily integrate others scripts with the playlists objects.
-- Data: json file for playlists is now formatted to be readable.
 ### Removed
 ### Fixed
+
+## [0.4.1] - 2021-06-13
+### Added
+- Integration: Listeners to share tracked playlist path with other panels (use 'window.NotifyOthers('Playlist manager: playlistPath', null)'). Used on [Playlist-Tools-SMP](https://github.com/regorxxx/Playlist-Tools-SMP) to use tracked playlist files as source for pools. i.e. Playlist A would match first a playlist within foobar with same name, then a playlist file with matching '#PLAYLIST:Playlist A' tag and finally any playlist file named 'Playlist A.m3u8'. Autoplaylists are excluded (use queries instead) and fpl files too. This feature allows to use virtual playlists as containers, where you can easily collect tracks (since Playlist Manager allows to send tracks directly to a file without loading it) to be used later on pools without polluting the UI with tons of dummy playlists.
+- Colors: added menu option to change standard text color, switchange between CUI\DUI color or custom one.
+### Changed
+- Portable: when properties are set for the first time, now use relative paths on profile folder for portable installations (>= 1.6).
+- Integration: Moved some functions to 'playlist_manager_helpers.js' to easily integrate others scripts with the playlists objects.
+- Data: json file for playlists is now formatted to be readable.
+- Colors: background color menu moved to 'Set custom colour...' submenu. Now all colour options reside in the same submenu.
+### Removed
+### Fixed
+- Colors: panel did not repaint with default color values in some cases after using 'Reset all to default' entry.
+- Avoid 2 possible crashes when playlist json files get corrupted. Warns about it whit a popup (to restore a backup).
 
 ## [0.4.0] - 2021-06-07
 ### Added
@@ -110,7 +121,8 @@
 ### Removed
 ### Fixed
 
-[Unreleased]: https://github.com/regorxxx/Playlist-Manager-SMP/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/regorxxx/Playlist-Manager-SMP/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/regorxxx/Playlist-Manager-SMP/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/regorxxx/Playlist-Manager-SMP/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/regorxxx/Playlist-Manager-SMP/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/regorxxx/Playlist-Manager-SMP/compare/v0.2.2...v0.3.0